MDEV-20612: Replace lock_sys.mutex with lock_sys.latch
For now, we will acquire the lock_sys.latch only in exclusive mode,
that is, use it as a mutex. This is preparation for the next commit
where we will introduce srw_lock shards for page_id_t, table_id_t and
actually acquire lock_sys.latch.rd_lock() for individual shards.